Always show help indicator on status bar
The help indicator ("Help - Ctrl+/") was previously hidden whenever
there was a status message. This made it invisible most of the time
since the status bar typically shows file info, cursor position, etc.

Changed to always show the help indicator on the bottom right of the
screen, regardless of status bar contents. It will only be hidden if
the terminal is too narrow (less than 15 characters wide).

This ensures users can always discover the help keybinding.